requirements.txt
lark==1.1.9
google-generativeai==0.5.4
python-dotenv==1.0.1
pandas==2.2.2
